# docs/swagger.yaml
basePath: /
definitions:
  models.Password:
    properties:
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      site:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple API for managing user passwords
  title: Password Manager API
  version: "1.0"
paths:
  /password:
    get:
      description: Retrieve passwords by phone number and site name
      parameters:
      - description: Phone number
        in: query
        name: phone
        required: true
        type: string
      - description: Site name
        in: query
        name: site
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of passwords matching criteria
          schema:
            items:
              $ref: '#/definitions/models.Password'
            type: array
        "400":
          description: Phone and site are required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch passwords
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Passwords by Site
      tags:
      - Password
    post:
      consumes:
      - application/json
      description: Create a new password
      parameters:
      - description: Password data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.Password'
      produces:
      - application/json
      responses:
        "201":
          description: Password created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Password
      tags:
      - Password
  /password/{phone}:
    get:
      description: Retrieve all passwords associated with a phone number
      parameters:
      - description: Phone number
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of passwords
          schema:
            items:
              $ref: '#/definitions/models.Password'
            type: array
        "400":
          description: Phone number is required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch passwords
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All Passwords by Phone
      tags:
      - Password
swagger: "2.0"